mysql/requirements.txt
streamlit
sqlalchemy
mysql-connector-python
pymysql
python-dotenv
cryptography
pandas